refactor: clean up helmut after initial refactoring (#544)
* refactor: remove redundant Logging class from helmut library

It creates unnecessary complexity by adding this Logging class to every class that wants to implement logging.

* refactor: get rid of metaclasses and abstract methods/properties

* refactor: remove helmut.manager.object and helmut.manager.jobs.job

* refactor: remove helmut.manager

* refactor: remove helmut.misc.collection

* refactor: remove helmut.manager.jobs.Jobs

* refactor: remove helmut.exceptions.UnsupportedConnectorError

* refactor: remove contype variable

* refactor: remove helmut.connector.base

* refactor: remove helmut.splunk.base

* refactor: remove deprecated username, password and args params

* style: run pre-commit

* refactor: remove deprecated namespace option from helmut.connector.sdk

* refactor: SDKConnector sharing default value
