+++ pytest_splunk_addon/helmut/__init__.py
@@ -13,11 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""
-@author: Nicklas Ansman-Giertz
-@contact: U{ngiertz@splunk.com<mailto:ngiertz@splunk.com>}
-@since: 2011-11-23
-"""
-from . import log
+from pytest_splunk_addon.helmut import log
 
 log.setup_logger(debug=True)
