+++ pytest_splunk_addon/helmut/connector/__init__.py
@@ -13,14 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""
-Module for handling generic connections with a Splunk instance.
-
-@author: Nicklas Ansman-Giertz
-@contact: U{ngiertz@splunk.com<mailto:ngiertz@splunk.com>}
-@since: 2011-11-21
-"""
-
 __all__ = ["sdk"]
 
 from .sdk import SDKConnector
