+++ pytest_splunk_addon/helmut/exceptions/__init__.py
@@ -13,19 +13,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-class UnsupportedConnectorError(BaseException):
-    """
-    Raised in manager_utils during creation of a Manager when the class type
-    of the new Manager object is being determined. When mapping the class type
-    of the connector currently in use to the appropriate and corresponding
-    subclass of Manager associated with that connector type, this
-    exception will be raised if the Manager has no subclasses associated with
-    that connector type.
-
-    Also raised during creation of a connector for a splunk instance if
-    attempting to create a connector type which isn't supported or doesnt exist
-    """
-
-    def __init__(self, message=None):
-        message = message or "The specified connector is not supported"
-        super(UnsupportedConnectorError, self).__init__(message)
