+++ pytest_splunk_addon/helmut/exceptions/job.py
@@ -13,23 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""
-@author: Nicklas Ansman-Giertz
-@contact: U{ngiertz@splunk.com<mailto:ngiertz@splunk.com>}
-@since: 2011-11-23
-"""
-from abc import ABCMeta
-
-from future.utils import with_metaclass
-
-from pytest_splunk_addon.helmut.log import Logging
-
-
-class ItemFromManager(with_metaclass(ABCMeta, Logging)):
-    def __init__(self, connector):
-        self._connector = connector
-        Logging.__init__(self)
+class JobNotFound(RuntimeError):
+    def __init__(self, sid):
+        self.sid = sid
+        super(JobNotFound, self).__init__(self._error_message)
 
     @property
-    def connector(self):
-        return self._connector
+    def _error_message(self):
+        return "Could not find a job with SID {sid}".format(sid=self.sid)
