+++ pytest_splunk_addon/helmut/manager/jobs/sdk/__init__.py
@@ -13,27 +13,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""
-This module is a specialized version of the search_manager module for the SDK
+import logging
 
-@author: Nicklas Ansman-Giertz
-@contact: U{ngiertz@splunk.com<mailto:ngiertz@splunk.com>}
-@since: 2011-11-23
-"""
-
-from pytest_splunk_addon.helmut.manager.jobs import Jobs, JobNotFound
+from pytest_splunk_addon.helmut.exceptions.job import JobNotFound
 from pytest_splunk_addon.helmut.manager.jobs.sdk.job import SDKJobWrapper
 
+LOGGER = logging.getLogger("helmut")
 
-class SDKJobsWrapper(Jobs):
+
+class SDKJobsWrapper:
+    def __init__(self, connector):
+        self._connector = connector
+
     @property
     def _service(self):
-        return self.connector.service
+        return self._connector.service
 
     def create(self, query, **kwargs):
-        self.logger.info("Creating job with query: %s" % query)
+        LOGGER.info("Creating job with query: %s" % query)
         job = self._service.jobs.create(query, **kwargs)
-        return SDKJobWrapper(self.connector, job)
+        return SDKJobWrapper(self._connector, job)
+
+    def __call__(self):
+        return list(self.items())
+
+    def __len__(self):
+        return len(list(self.items()))
+
+    def __iter__(self):
+        for item in list(self.items()):
+            yield item
 
     def __contains__(self, sid):
         for job in self:
@@ -47,8 +56,6 @@ class SDKJobsWrapper(Jobs):
                 return job
         raise JobNotFound(sid)
 
-    # Required from Collection
-
     def items(self):
         jobs = self._service.jobs
-        return [SDKJobWrapper(self.connector, job) for job in jobs]
+        return [SDKJobWrapper(self._connector, job) for job in jobs]
